+++ cmd/bricksllm/main.go
@@ -96,6 +96,11 @@ func main() {
 		log.Sugar().Fatalf("error create create at index for keys: %v", err)
 	}
 
+	err = store.CreateKeyIndexForKeys()
+	if err != nil {
+		log.Sugar().Fatalf("error create create at index for keys: %v", err)
+	}
+
 	err = store.CreateEventsTable()
 	if err != nil {
 		log.Sugar().Fatalf("error creating events table: %v", err)
@@ -155,18 +160,6 @@ func main() {
 	if err != nil {
 		log.Sugar().Fatalf("error creating user id for users table: %v", err)
 	}
-
-	memStore, err := memdb.NewMemDb(store, log, cfg.InMemoryDbUpdateInterval)
-	if err != nil {
-		log.Sugar().Fatalf("cannot initialize memdb: %v", err)
-	}
-	memStore.Listen()
-
-	psMemStore, err := memdb.NewProviderSettingsMemDb(store, log, cfg.InMemoryDbUpdateInterval)
-	if err != nil {
-		log.Sugar().Fatalf("cannot initialize provider settings memdb: %v", err)
-	}
-	psMemStore.Listen()
 
 	cpMemStore, err := memdb.NewCustomProvidersMemDb(store, log, cfg.InMemoryDbUpdateInterval)
 	if err != nil {
@@ -287,6 +280,30 @@ func main() {
 		log.Sugar().Fatalf("error connecting to user access redis storage: %v", err)
 	}
 
+	providerSettingsRedisCache := redis.NewClient(&redis.Options{
+		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
+		Password: cfg.RedisPassword,
+		DB:       9,
+	})
+
+	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+	if err := costRedisStorage.Ping(ctx).Err(); err != nil {
+		log.Sugar().Fatalf("error connecting to provider settings redis storage: %v", err)
+	}
+
+	keysRedisCache := redis.NewClient(&redis.Options{
+		Addr:     fmt.Sprintf("%s:%s", cfg.RedisHosts, cfg.RedisPort),
+		Password: cfg.RedisPassword,
+		DB:       10,
+	})
+
+	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+	if err := costRedisStorage.Ping(ctx).Err(); err != nil {
+		log.Sugar().Fatalf("error connecting to keys redis storage: %v", err)
+	}
+
 	rateLimitCache := redisStorage.NewCache(rateLimitRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 	costLimitCache := redisStorage.NewCache(costLimitRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 	costStorage := redisStorage.NewStore(costRedisStorage, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
@@ -298,13 +315,16 @@ func main() {
 	userCostStorage := redisStorage.NewStore(userCostRedisStorage, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 	userAccessCache := redisStorage.NewAccessCache(userAccessRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
 
+	psCache := redisStorage.NewProviderSettingsCache(providerSettingsRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
+	keysCache := redisStorage.NewKeysCache(keysRedisCache, cfg.RedisWriteTimeout, cfg.RedisReadTimeout)
+
 	v := validator.NewValidator(costLimitCache, rateLimitCache, costStorage)
 
-	m := manager.NewManager(store, costLimitCache, rateLimitCache, accessCache)
+	m := manager.NewManager(store, costLimitCache, rateLimitCache, accessCache, keysCache)
 	krm := manager.NewReportingManager(costStorage, store, store, v)
-	psm := manager.NewProviderSettingsManager(store, psMemStore)
+	psm := manager.NewProviderSettingsManager(store, psCache)
 	cpm := manager.NewCustomProvidersManager(store, cpMemStore)
-	rm := manager.NewRouteManager(store, store, rMemStore, psMemStore)
+	rm := manager.NewRouteManager(store, store, rMemStore, psm)
 	pm := manager.NewPolicyManager(store, rMemStore)
 	um := manager.NewUserManager(store, store)
 
@@ -339,7 +359,7 @@ func main() {
 
 	rec := recorder.NewRecorder(costStorage, userCostStorage, costLimitCache, userCostLimitCache, ce, store)
 	rlm := manager.NewRateLimitManager(rateLimitCache, userRateLimitCache)
-	a := auth.NewAuthenticator(psm, memStore, rm, store)
+	a := auth.NewAuthenticator(psm, m, rm, store)
 
 	c := cache.NewCache(apiCache)
 
@@ -360,7 +380,7 @@ func main() {
 	scanner := pii.NewScanner(detector)
 	cd := custompolicy.NewOpenAiDetector(cfg.CustomPolicyDetectionTimeout, cfg.OpenAiApiKey)
 
-	ps, err := proxy.NewProxyServer(log, *modePtr, *privacyPtr, c, m, rm, a, psm, cpm, store, memStore, ce, ace, aoe, v, rec, messageBus, rlm, cfg.ProxyTimeout, accessCache, userAccessCache, pm, scanner, cd, die, um)
+	ps, err := proxy.NewProxyServer(log, *modePtr, *privacyPtr, c, m, rm, a, psm, cpm, store, ce, ace, aoe, v, rec, messageBus, rlm, cfg.ProxyTimeout, accessCache, userAccessCache, pm, scanner, cd, die, um, cfg.RemoveUserAgent)
 	if err != nil {
 		log.Sugar().Fatalf("error creating proxy http server: %v", err)
 	}
@@ -372,8 +392,6 @@ func main() {
 	<-quit
 
 	eventConsumer.Stop()
-	memStore.Stop()
-	psMemStore.Stop()
 	cpMemStore.Stop()
 	rMemStore.Stop()